feat : 아이템 상호작용 기능 추가 / 월드맵 탐험도에 따라 플레이어 능력치가 상승하도록 수정

+++ Assets/Scripts/Weapon/Proximity/WeaponStick.cs
@@ -19,7 +19,7 @@ public class WeaponStick : MonoBehaviour
     private BoxCollider _boxCollider;
     private Coroutine _attackCoroutine;
     private List<ITakeDamage> _hitTargets = new List<ITakeDamage>();
-    private int _damage = 0;
+    private float _damage = 0;
 
     void Awake()
     {
